refactor: user tests
